requirements.txt
scrapy
django
south
django-model-utils
django-reversion
django-xadmin
apscheduler
fabric
djangorestframework
pytz